area/ui: Add a search input for easy filtering of nodes
