+++ ui/packages/shared/components/src/Dropdown/index.tsx
@@ -1,18 +1,18 @@
 import {Menu, Transition} from '@headlessui/react';
 import React, {Fragment} from 'react';
 import {ChevronDownIcon} from '@heroicons/react/solid';
-import Button, {ButtonColor} from '../Button';
+import Button from '../Button';
 import cx from 'classnames';
 
 const Dropdown = ({
   text,
   element,
-  color = 'neutral',
+  variant = 'neutral',
   children,
 }: {
   text: string;
   element?: React.ReactNode;
-  color?: ButtonColor;
+  variant?: 'neutral' | 'link' | 'primary' | undefined;
   children?: React.ReactNode;
 }) => {
   return (
@@ -21,7 +21,7 @@ const Dropdown = ({
         <div>
           <Menu.Button as="div">
             {element !== undefined ? (
-              <Button color={color}>
+              <Button variant={variant}>
                 {text}{' '}
                 <ChevronDownIcon
                   className="w-5 h-5 ml-2 -mr-1 text-violet-200 hover:text-violet-100"
