+++ ui/packages/shared/components/src/index.tsx
@@ -1,4 +1,4 @@
-import Button, {ButtonColor} from './Button';
+import Button from './Button';
 import ButtonGroup from './ButtonGroup';
 import Card from './Card';
 import DateTimeRangePicker, {DateTimeRange} from './DateTimeRangePicker';
@@ -6,6 +6,7 @@ import DateTimePicker from './DateTimePicker';
 import Dropdown from './Dropdown';
 import GrpcMetadataContext, {GrpcMetadataProvider, useGrpcMetadata} from './GrpcMetadataContext';
 import Input from './Input';
+import FlamegraphTooltip from './FlamegraphTooltip';
 import MatchersInput from './MatchersInput';
 import MetricsGraph from './MetricsGraph';
 import ParcaThemeContext, {ParcaThemeProvider, useParcaTheme} from './ParcaThemeContext';
@@ -15,11 +16,12 @@ import ProfileMetricsGraph from './ProfileMetricsGraph';
 import ProfileSelector from './ProfileSelector';
 import Select from './Select';
 import type {SelectElement} from './Select';
+import SearchNodes from './SearchNodes';
 import Spinner from './Spinner';
 import Tab from './Tab';
 import EmptyState from './EmptyState';
 
-export type {ButtonColor, PillVariant, SelectElement};
+export type {PillVariant, SelectElement};
 
 export {
   Button,
@@ -29,6 +31,7 @@ export {
   DateTimeRange,
   DateTimeRangePicker,
   Dropdown,
+  FlamegraphTooltip,
   GrpcMetadataContext,
   GrpcMetadataProvider,
   Input,
@@ -41,6 +44,7 @@ export {
   ProfileMetricsGraph,
   ProfileSelector,
   Select,
+  SearchNodes,
   Spinner,
   Tab,
   EmptyState,
