+++ ui/packages/shared/profile/src/IcicleGraph.tsx
@@ -3,9 +3,9 @@ import {throttle} from 'lodash';
 import {pointer} from 'd3-selection';
 import {scaleLinear} from 'd3-scale';
 import {Flamegraph, FlamegraphNode, FlamegraphRootNode} from '@parca/client';
-import {usePopper} from 'react-popper';
-import {getLastItem, valueFormatter, diffColor} from '@parca/functions';
-import {useAppSelector, selectDarkMode} from '@parca/store';
+import {FlamegraphTooltip} from '@parca/components';
+import {getLastItem, diffColor} from '@parca/functions';
+import {useAppSelector, selectDarkMode, selectSearchNodeString} from '@parca/store';
 
 interface IcicleGraphProps {
   graph: Flamegraph;
@@ -27,15 +27,6 @@ interface IcicleGraphNodesProps {
   setHoveringNode: (node: FlamegraphNode | FlamegraphRootNode | undefined) => void;
   path: string[];
   xScale: (value: number) => number;
-}
-
-interface FlamegraphTooltipProps {
-  x: number;
-  y: number;
-  unit: string;
-  total: number;
-  hoveringNode: FlamegraphNode | FlamegraphRootNode | undefined;
-  contextElement: Element | null;
 }
 
 interface IcicleGraphRootNodeProps {
@@ -85,8 +76,14 @@ function IcicleRect({
   onClick,
   curPath,
 }: IcicleRectProps) {
+  const currentSearchString = useAppSelector(selectSearchNodeString);
   const isFaded = curPath.length > 0 && name !== curPath[curPath.length - 1];
   const styles = isFaded ? fadedIcicleRectStyles : icicleRectStyles;
+
+  const isSearchMatch =
+    currentSearchString !== undefined &&
+    currentSearchString !== '' &&
+    name.includes(currentSearchString);
 
   return (
     <g
@@ -102,7 +99,7 @@ function IcicleRect({
         width={width - 1}
         height={height - 1}
         style={{
-          fill: color,
+          fill: isSearchMatch ? '#e39c9c' : color,
         }}
       />
       {width > 5 && (
@@ -231,196 +228,6 @@ export function IcicleGraphNodes({
 }
 
 const MemoizedIcicleGraphNodes = React.memo(IcicleGraphNodes);
-
-const FlamegraphNodeTooltipTableRows = ({
-  hoveringNode,
-}: {
-  hoveringNode: FlamegraphNode;
-}): JSX.Element => {
-  if (hoveringNode.meta === undefined) return <></>;
-
-  return (
-    <>
-      {hoveringNode.meta.pb_function?.filename !== undefined &&
-        hoveringNode.meta.pb_function?.filename !== '' && (
-          <tr>
-            <td className="w-1/5">File</td>
-            <td className="w-4/5">
-              {hoveringNode.meta.pb_function.filename}
-              {hoveringNode.meta.line?.line !== undefined && hoveringNode.meta.line?.line !== 0
-                ? ` +${hoveringNode.meta.line.line.toString()}`
-                : `${
-                    hoveringNode.meta.pb_function?.startLine !== undefined &&
-                    hoveringNode.meta.pb_function?.startLine !== 0
-                      ? ` +${hoveringNode.meta.pb_function.startLine.toString()}`
-                      : ''
-                  }`}
-            </td>
-          </tr>
-        )}
-      {hoveringNode.meta.location?.address !== undefined &&
-        hoveringNode.meta.location?.address !== 0 && (
-          <tr>
-            <td className="w-1/5">Address</td>
-            <td className="w-4/5">{' 0x' + hoveringNode.meta.location.address.toString(16)}</td>
-          </tr>
-        )}
-      {hoveringNode.meta.mapping !== undefined && hoveringNode.meta.mapping.file !== '' && (
-        <tr>
-          <td className="w-1/5">Binary</td>
-          <td className="w-4/5">{getLastItem(hoveringNode.meta.mapping.file)}</td>
-        </tr>
-      )}
-    </>
-  );
-};
-
-function generateGetBoundingClientRect(contextElement: Element, x = 0, y = 0) {
-  const domRect = contextElement.getBoundingClientRect();
-  return () =>
-    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-    ({
-      width: 0,
-      height: 0,
-      top: domRect.y + y,
-      left: domRect.x + x,
-      right: domRect.x + x,
-      bottom: domRect.y + y,
-    } as ClientRect);
-}
-
-const virtualElement = {
-  getBoundingClientRect: () =>
-    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-    ({
-      width: 0,
-      height: 0,
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-    } as ClientRect),
-};
-
-export const FlamegraphTooltip = ({
-  x,
-  y,
-  unit,
-  total,
-  hoveringNode,
-  contextElement,
-}: FlamegraphTooltipProps): JSX.Element => {
-  const [popperElement, setPopperElement] = useState<HTMLDivElement | null>(null);
-
-  const {styles, attributes, ...popperProps} = usePopper(virtualElement, popperElement, {
-    placement: 'auto-start',
-    strategy: 'absolute',
-    modifiers: [
-      {
-        name: 'preventOverflow',
-        options: {
-          tether: false,
-          altAxis: true,
-        },
-      },
-      {
-        name: 'offset',
-        options: {
-          offset: [30, 30],
-        },
-      },
-    ],
-  });
-
-  const update = popperProps.update;
-
-  useEffect(() => {
-    if (contextElement != null) {
-      virtualElement.getBoundingClientRect = generateGetBoundingClientRect(contextElement, x, y);
-      update?.();
-    }
-  }, [x, y, contextElement, update]);
-
-  if (hoveringNode === undefined || hoveringNode == null) return <></>;
-
-  const hoveringNodeCumulative = parseFloat(hoveringNode.cumulative);
-  const diff = hoveringNode.diff === undefined ? 0 : parseFloat(hoveringNode.diff);
-  const prevValue = hoveringNodeCumulative - diff;
-  const diffRatio = Math.abs(diff) > 0 ? diff / prevValue : 0;
-  const diffSign = diff > 0 ? '+' : '';
-  const diffValueText = diffSign + valueFormatter(diff, unit, 1);
-  const diffPercentageText = diffSign + (diffRatio * 100).toFixed(2) + '%';
-  const diffText = `${diffValueText} (${diffPercentageText})`;
-
-  const hoveringFlamegraphNode = hoveringNode as FlamegraphNode;
-  const metaRows =
-    hoveringFlamegraphNode.meta === undefined ? (
-      <></>
-    ) : (
-      <FlamegraphNodeTooltipTableRows hoveringNode={hoveringNode as FlamegraphNode} />
-    );
-
-  return (
-    <div ref={setPopperElement} style={styles.popper} {...attributes.popper}>
-      <div className="flex">
-        <div className="m-auto">
-          <div
-            className="border-gray-300 dark:border-gray-500 bg-gray-50 dark:bg-gray-900 rounded-lg p-3 shadow-lg opacity-90"
-            style={{borderWidth: 1}}
-          >
-            <div className="flex flex-row">
-              <div className="ml-2 mr-6">
-                <span className="font-semibold">
-                  {hoveringFlamegraphNode.meta === undefined ? (
-                    <p>root</p>
-                  ) : (
-                    <>
-                      {hoveringFlamegraphNode.meta.function !== undefined &&
-                      hoveringFlamegraphNode.meta.function.name !== '' ? (
-                        <p>{hoveringFlamegraphNode.meta.function.name}</p>
-                      ) : (
-                        <>
-                          {hoveringFlamegraphNode.meta.location !== undefined &&
-                          parseInt(hoveringFlamegraphNode.meta.location.address, 10) !== 0 ? (
-                            <p>
-                              {'0x' + hoveringFlamegraphNode.meta.location.address.toString(16)}
-                            </p>
-                          ) : (
-                            <p>unknown</p>
-                          )}
-                        </>
-                      )}
-                    </>
-                  )}
-                </span>
-                <span className="text-gray-700 dark:text-gray-300 my-2">
-                  <table className="table-fixed">
-                    <tbody>
-                      <tr>
-                        <td className="w-1/5">Cumulative</td>
-                        <td className="w-4/5">
-                          {valueFormatter(hoveringNodeCumulative, unit, 2)} (
-                          {((hoveringNodeCumulative * 100) / total).toFixed(2)}%)
-                        </td>
-                      </tr>
-                      {hoveringNode.diff !== undefined && diff !== 0 && (
-                        <tr>
-                          <td className="w-1/5">Diff</td>
-                          <td className="w-4/5">{diffText}</td>
-                        </tr>
-                      )}
-                      {metaRows}
-                    </tbody>
-                  </table>
-                </span>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 export function IcicleGraphRootNode({
   node,
