+++ ui/packages/shared/store/src/slices/profileSlice.ts
@@ -4,11 +4,13 @@ import type {RootState} from '../store';
 // Define a type for the slice state
 interface ProfileState {
   compare: boolean;
+  searchNodeString: string | undefined;
 }
 
 // Define the initial state using that type
 const initialState: ProfileState = {
   compare: false,
+  searchNodeString: undefined,
 };
 
 export const profileSlice = createSlice({
@@ -19,12 +21,16 @@ export const profileSlice = createSlice({
     setCompare: (state, action: PayloadAction<boolean>) => {
       state.compare = action.payload;
     },
+    setSearchNodeString: (state, action: PayloadAction<string>) => {
+      state.searchNodeString = action.payload;
+    },
   },
 });
 
-export const {setCompare} = profileSlice.actions;
+export const {setCompare, setSearchNodeString} = profileSlice.actions;
 
 // Other code such as selectors can use the imported `RootState` type
 export const selectCompareMode = (state: RootState) => state.profile.compare;
+export const selectSearchNodeString = (state: RootState) => state.profile.searchNodeString;
 
 export default profileSlice.reducer;
